Consolidate utility classes and functions into `zigpy_znp.utils`
